Use one file for rendering, server and client side.
